Move CameraCapability to the Camera package
